Bugfix: MatchError for FeatureExtractorDeliveryMatches
The `scala.MatchError` turns out to be the result of assigning a wrong
type during the creation of output dataframe.

After the test run, LR produces results of the following (a bit better
    than before).

```
(20161005,(0.7205119976029932,(0.7156838997309312,0.834350958603004,0.770474992335423,(437280,173716,267942,86816))))
(20161006,(0.7164570307343775,(0.7092203765631757,0.8063604631835559,0.7546773571514223,(465728,190948,320365,111840))))
(20161007,(0.7211083902737258,(0.7163749142070429,0.8136285362853628,0.7619107952864342,(529184,209513,354586,121216))))
(20161008,(0.7354851996277948,(0.6750460233048189,0.8391729476471786,0.7482145872443557,(527296,253830,435546,101056))))
(20161009,(0.7572515732878264,(0.6240444650847358,0.80747806180847,0.7040087818143261,(338624,204004,492315,80736))))
```

+++ src/main/scala/PredictorLR.scala
@@ -29,7 +29,7 @@ object PredictorLR {
         val array = r.split('(')(1).split(')')(0).split(", ")
         val activeness = array(0)
         val features = array.slice(1, array.size).map(_.toDouble)
-        (date.toInt, LabeledPoint(activeness.toInt, Vectors.dense(features)))
+        (date.toInt, LabeledPoint(activeness.toDouble.toInt, Vectors.dense(features)))
       })
 
     val dividerDate = 20161005
